refactor(UiMapping): ui 모델로 객체를 변경하는 로직을 이동
- UiMapping -> 각 ui 모델 클래스의 최상위
